Compiler now sweeps for opcode arrays with multiple references
Upon encountering a whitelisted op_array that has multiple references
but does _not_ have an enforcement op_array, we now do something
intelligent depending on the situation (this previously sometimes
caused a crash). In this event, we check to see if our current
compilation sweep would encounter every other reference to the
op_array in question. If so, we can safely compile the doc annotation
and free the opcode array's corresponding literal array, because
we know that all other references will be later fixed up on this
sweep (this commonly happens when compiling files containing a
class hierarchy). If not, we throw a notice to the user and bail
before compilation so that we don't corrupt the other references.